+++ src-bin/clean.ml
@@ -8,17 +8,27 @@ open Bos_setup
 
 (* Command *)
 
-let clean () pkg_file build_dir name =
+let clean_args name build_dir =
+  let on_some_use_opt opt to_arg = function
+  | None -> Cmd.empty
+  | Some value -> Cmd.(v opt % to_arg value)
+  in
+  let verb = Cli.propagate_verbosity_to_pkg_file () in
+  let build_dir = on_some_use_opt "--build-dir" Cmd.p build_dir in
+  let name = on_some_use_opt "--pkg-name" (fun n -> n) name in
+  Cmd.(verb %% name %% build_dir)
+
+let clean () pkg_file name build_dir =
+  let pkg = Topkg_care.Pkg.v ?build_dir ?name pkg_file in
+  let args = clean_args name build_dir in
+  let out = OS.Cmd.out_stdout in
   begin
-    let p = Topkg_care.Pkg.v ?build_dir ?name pkg_file in
-    Topkg_care.Pkg.name p
-    >>= fun name -> Fpath.of_string (name ^ ".install")
-    >>= fun install -> OS.File.delete install
-    >>= fun () -> Topkg_care.Pkg.build_dir p
-    >>= fun build_dir -> OS.Dir.delete ~recurse:true build_dir
-    >>= fun () -> Ok 0
+    OS.Dir.current ()
+    >>= fun dir -> Topkg_care.Pkg.clean pkg ~dir ~args ~out
+    >>| (function ((), (_, `Exited 0)) -> 0 | _ -> 1)
   end
   |> Cli.handle_error
+
 
 (* Command line interface *)
 
@@ -26,17 +36,20 @@ open Cmdliner
 
 let doc = "Clean the package's build"
 let man =
-  [ `S "DESCRIPTION";
-    `P "The $(b,$(tname)) command deletes the package's build directory
-        and its OPAM install file."
+  [ `S "SYNOPSIS";
+     `P "$(b,$(mname)) $(b,$(tname)) [$(i,OPTION)]...";
+    `S "DESCRIPTION";
+    `P "The $(b,$(tname)) command deletes the package's build and its OPAM
+        install file. This is equivalent to invoke:";
+    `Pre "ocaml ./pkg/pkg.ml clean";
   ] @ Cli.common_opts_man @ [
     `S "ENVIRONMENT VARIABLES";
   ] @ Cli.see_also ~cmds:["topkg-build"]
 
 let cmd =
   let info = Term.info "clean" ~sdocs:Cli.common_opts ~doc ~man in
-  let t = Term.(pure clean $ Cli.setup $ Cli.pkg_file $ Cli.build_dir $
-                Cli.pkg_name)
+  let t = Term.(pure clean $ Cli.setup $ Cli.pkg_file $ Cli.pkg_name $
+                Cli.build_dir)
   in
   (t, info)