+++ src-care/topkg_care.mli
@@ -440,6 +440,12 @@ module Pkg : sig
     t -> dir:Fpath.t -> args:Cmd.t ->
     out:(OS.Cmd.run_out -> ('a, R.msg) result) -> ('a, R.msg) result
 
+  (** {1 Clean} *)
+
+  val clean :
+    t -> dir:Fpath.t -> args:Cmd.t ->
+    out:(OS.Cmd.run_out -> ('a, R.msg) result) -> ('a, R.msg) result
+
   (** {1 Distribution} *)
 
   val distrib_archive : t -> keep_dir:bool -> (Fpath.t, R.msg) result