+++ src/topkg_build.mli
@@ -17,7 +17,9 @@ val v :
   ?dir:Topkg_fpath.t ->
   ?pre:(Topkg_conf.t -> unit result) ->
   ?cmd:(Topkg_conf.t -> Topkg_conf.os -> Topkg_fpath.t list -> unit result) ->
-  ?post:(Topkg_conf.t -> unit result) -> unit -> t
+  ?post:(Topkg_conf.t -> unit result) ->
+  ?clean:(Topkg_conf.os -> build_dir:Topkg_fpath.t -> unit result) ->
+  unit -> t
 
 val with_dir : t -> Topkg_fpath.t -> t
 
@@ -28,6 +30,8 @@ val cmd :
   t -> (Topkg_conf.t -> Topkg_conf.os -> Topkg_fpath.t list -> unit result)
 
 val post : t -> (Topkg_conf.t -> unit result)
+val clean : t -> (Topkg_conf.os -> build_dir:Topkg_fpath.t -> unit result)
+
 val codec : t Topkg_codec.t
 
 (*---------------------------------------------------------------------------